Fix compilation when TRUST configured with -with-64-bit-indices

+++ src/Kernel/ICoCo/Convert_ICoCoTrioField.cpp
@@ -1,5 +1,5 @@
 /****************************************************************************
-* Copyright (c) 2021, CEA
+* Copyright (c) 2022, CEA
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
@@ -90,14 +90,14 @@ void build_triofield(const Champ_Generique_base& ch, ICoCo::TrioField& afield)
     {
       const IntTab& conn = loc_faces ? zvf.face_sommets() : zvf.zone().les_elems();
       //le seul moyen qu'on a d'eviter que des polygones soient pris pour des quadrilateres est d'avoir un tableau de connectivite de largeur > 4...
-      afield._nodes_per_elem = std::max(conn.dimension(1), type_elem == "POLYGONE" || type_elem == "POLYGONE_3D"  || type_elem == "POLYEDRE"  ? 5 : 0);
+      afield._nodes_per_elem = std::max(conn.dimension(1), type_elem == "POLYGONE" || type_elem == "POLYGONE_3D"  || type_elem == "POLYEDRE"  ? (int) 5 : 0);
       afield._connectivity = new int[afield._nb_elems * afield._nodes_per_elem];
       for (int i = 0; i < afield._nb_elems; i++) for (int j = 0; j < afield._nodes_per_elem; j++)
           afield._connectivity[afield._nodes_per_elem * i + j] = j < conn.dimension(1) ? conn(i, j) : -1;
     }
   else //maillage de polyedres -> connectivite au format MEDCoupling, a faire a la main
     {
-      afield._nodes_per_elem = std::max(zvf.elem_faces().dimension(1) * (zvf.face_sommets().dimension(1) + 1), 9); //un -1 apres chaque face : au moins 9 pour eviter un papillonage
+      afield._nodes_per_elem = std::max(zvf.elem_faces().dimension(1) * (zvf.face_sommets().dimension(1) + 1), (int) 9); //un -1 apres chaque face : au moins 9 pour eviter un papillonage
       int *p = afield._connectivity = new int[afield._nb_elems * afield._nodes_per_elem];
       for (int e = 0, f, s, i, j; e < afield._nb_elems; e++)
         {
